la blinde d'erreurs de compil, beaucoup de correction deja

+++ src/lib_eso.c
@@ -1,7 +1,8 @@
 // eso = eval stack operation
+#include "../lib/lib_eso.h"
+
 // stands for the operation 
 // no integer, only double, why ? because you can add 1233.32 and 22.1 => 1255.33
-#include <stdint.h>
 
 // add +
 double add(double nbrA, double nbrB){
@@ -25,7 +26,7 @@ double div(double nbrA, double nbrB){
 
 // euclidian %
 double mod(double nbrA, double nbrB){
-	return nbrA % nbrB;
+	return (double)((int)nbrA % (int)nbrB);
 }
 
 // power
